tools/ceph_objectstore_tool: add op "expand-log"
There are circumstances in which, although we can't find enough OSDs to
form a valid ec acting set, we can enough OSDs that are fully backfilled
and it's only because their PGLog are lagged behind that they can't be
added the acting set. In this case, right now, we can only drop the
entire pg's data by recreating it.

This commit add the ability to fill in the pg log gap between the
backfilled OSDs and the auth log shard, so the pg can be brought to
active losing only part of its data and those uncorruped data can be
recovered.

Signed-off-by: Xuehan Xu <xuxuehan@qianxin.com>